requirements.txt
aiofiles==23.2.1
aiogram==3.15.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
attrs==23.2.0
certifi==2023.11.17
coloredlogs==15.0.1
executor==23.2
fasteners==0.19
frozenlist==1.4.1
humanfriendly==10.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pillow==10.2.0
property-manager==3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.9.0
verboselogs==1.7
yarl==1.9.4
redis
asyncpg